Refactor jump list implementation
Add more tests for jumplists

Inhibit jump hooks when saving `point' on the first jump backward

On the first jump backward, `point' is saved to the jumplist so the jumplist
needs to be rotated once backward before the real backward jump the user
requested. Prevent the hooks from running when saving `point' is useful if the
hooks update some visual elements of the window, for example highlighting the
line after a jump. In the case of the first backward jump, the hook would be
run twice if they are not inhibited.

Move `etags` jump integration to `evil-integration.el`

Add `evil-swap-out-all-jump-markers`

Jump markers that point to the `current-buffer` can exist in every window of
every frame. When the `current-buffer` is killed, its jump markers need to be
swapped in all of those windows, otherwise they become invalid and are silently
removed from the jumplists.

* Replace uses of `evil-swap-out-jump-markers` with
  `evil-swap-out-all-jump-markers`

Remove `evil-jumps-buffer-targets`, add `evil-jumps-allowed-buffer-patterns`

* Since `evil-jumps-ignored-file-patterns` is already a list of regular
  expressions, use a list of regular expressions for buffer targets as well for consistency.

Don't drop jumps that cause errors in `evil-loop-over-jumps`

Simplify `evil-swap-out-jump-markers`

* Since we cannot swap out the markers of a buffer target, don't try to.

Fix matching against buffer targets

Update `evil-jumps` commentary

Rename `evil--rotate-jumps-unless` to `evil-rotate-jumplist`

Also change the ring argument to a jumplist argument for
consistency with other jumplist operations.

Call `evil-maybe-copy-jumplist` in all jumplist modifying functions

Fix documentation of `evil-get-jumplist`

Use `with-demoted-errors` in `evil-loop-over-jumps`

Update documentation of evil-jumps.el

Update commentary and header of evil-jumps.el

Reduce access of implementation details

evil--do-jump: Be clearer when performing a jump

evil-get-jumplist: Be clearer

evil-jumplist-p: ringp -> ring-p